bis698_capstone-main/requirements.txt
altgraph==0.17.4
Babel==2.13.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.7
Jinja2==3.1.2
lxml==4.9.3
macholib==1.16.3
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
packaging==23.2
Pillow==10.0.1
protobuf==4.21.12
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
python-docx==0.8.11
reportlab==4.0.5
six==1.16.0
tkcalendar==1.6.1
tqdm==4.66.1
Werkzeug==2.3.7
